Adds correct puzzle solution
There is a more efficient way to solve this problem, however, I wanted
to changed as little of the part 1 answer as possible. This update for
part 2 only adds to the solution. For part 1, I optimized it in such a
way as to make performing other operations possible (not only "mul")
since I did not know what part 2 would be.

Answer submitted:
112272912

Response:
That's the right answer! You are one gold star closer to finding the Chief Historian.

To the King of the ages, immortal, invisible, the only God, be honor and glory forever and ever. Amen.